% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{os_example_image}
\alias{os_example_image}
\title{Example images provided as part of the opensportml package}
\usage{
os_example_image(choices = "2019 Wimbledon")
}
\arguments{
\item{choices}{string: which image files to return?
\itemize{
\item "2019 Wimbledon" - an image from the 2019 Wimbledon Men's Singles Championships, obtained from \url{https://en.wikipedia.org/wiki/2019_Wimbledon_Championships_\%E2\%80\%93_Men\%27s_Singles#/media/File:Wimbledon_2019_Nadal_v_Sousa.jpg} (CC BY-SA 4.0 license)
}}
}
\value{
Path to the image files
}
\description{
Example images provided as part of the opensportml package
}
