% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{os_ggcourt}
\alias{os_ggcourt}
\title{Minimal ggplot court plot}
\usage{
os_ggcourt(court = "tennis", line_colour = "black", ...)
}
\arguments{
\item{court}{string: court to plot, currently only "tennis"}

\item{line_colour}{string: colour to use for court lines}

\item{...}{: additional arguments, currently ignored}
}
\value{
A list that can be added to a \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}
}
\description{
Minimal ggplot court plot
}
\examples{

library(ggplot2)
ggplot() + os_ggcourt("tennis")

}
