% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{os_overlay_data}
\alias{os_overlay_data}
\title{Generate data suitable for creating a court overlay plot}
\usage{
os_overlay_data(court = "tennis", space = "court", court_ref, crop = TRUE)
}
\arguments{
\item{court}{string: court to plot, currently only "tennis"}

\item{space}{string: if "court", the data will be in court coordinates. If "image", the data will be transformed to image coordinates via \code{\link[ovideo:ov_transform_points]{ovideo::ov_transform_points()}}}

\item{court_ref}{data.frame: as returned by \code{\link[=os_shiny_court_ref]{os_shiny_court_ref()}}. Only required if \code{space} is "image"}

\item{crop}{logical: if \code{space} is "image", and \code{crop} is TRUE, the data will be cropped to the c(0, 1, 0, 1) bounding box (i.e. the limits of the image, in normalized coordinates). Requires that the \code{sf} package be installed}
}
\value{
A list of data.frames
}
\description{
Generate data suitable for creating a court overlay plot
}
